requirements.txt
click==6.6
requests==2.6.0
python-magic==0.4.6
